fix: 카카오 소셜로그인 실패시 카카오 인증 파기

+++ backend/src/main/java/com/bob/smash/dto/PaymentDTO.java
@@ -19,7 +19,7 @@ public class PaymentDTO {
 
     private String memberEmail;     // Member.emailId
     private String partnerBno;      // PartnerInfo.bno
-    private Long estimateIdx;       // Estimate.idx
+    private Integer estimateIdx;       // Estimate.idx
 
     private Integer price;
     private Status status;
